feat: include attempts in subsection attempt event
