+++ eox_nelp/edxapp_wrapper/test_backends/course_blocks_m_v1.py
@@ -1,0 +1,10 @@
+"""Test backend for course_blocks module."""
+from mock import Mock
+
+
+def get_student_module_as_dict_method():
+    """Return test function.
+    Returns:
+        Mock class.
+    """
+    return Mock()
